milestone3/requirements.txt
langchain>=0.1.17
langchain-core>=0.1.52
langchain-community>=0.0.38
langchain-openai>=0.1.6
faiss-cpu
python-dotenv
